🚚 Rename files to fit the component name
