+++ src/app/components/SummaryHighlightItem/SummaryHighlightItem.examples.tsx
@@ -1,5 +1,5 @@
 import { Example } from '../examples';
-import SummaryHighlightItem from './SummaryHighlight';
+import SummaryHighlightItem from './SummaryHighlightItem';
 
 export const LolSummaryHighlightItem: Example = () => (
   <SummaryHighlightItem highlights={['test, test']} timestamp="1627390756" />
