+++ src/app/components/examples.ts
@@ -9,4 +9,4 @@ export * from './MatchItem/MatchItem.examples';
 export * from './Upload/Upload.examples';
 export * from './User/User.example';
 export * from './VideoHighlight/VideoHighlight.examples';
-export * from './SummaryHighlight/SummaryHighlight.examples';
+export * from './SummaryHighlightItem/SummaryHighlightItem.examples';
